+++ lib/flow/spec_helper.rb
@@ -1,7 +1,8 @@
 # frozen_string_literal: true
 
-require "instructor/spec_helper"
+require "spicerack/spec_helper"
 
-require_relative "custom_matchers"
-require_relative "shoulda_matcher_helper"
-require_relative "rspec_configuration"
+require_relative "spec_helper/custom_matchers"
+require_relative "spec_helper/shared_contexts"
+require_relative "spec_helper/shoulda_matcher_helper"
+require_relative "spec_helper/rspec_configuration"