+++ spec/support/shared_context/with_an_example_state.rb
@@ -3,7 +3,7 @@
 RSpec.shared_context "with an example state" do |state_modules = nil|
   subject(:example_state) { example_state_class.new(*input) }
 
-  let(:root_state_class) { Class.new(Instructor::Base) }
+  let(:root_state_class) { Class.new(Spicerack::InputModel) }
   let(:example_state_class) do
     root_state_class.tap do |state_class|
       Array.wrap(state_modules).each { |state_module| state_class.include state_module }